create issue #6: Document the behaviour pertaining to corrupted MathType equation
This issue is likely not a big in Spire.Doc as the behaviour is reproducible with the Office Word application.
That is, manually saving the document in question as .docx using Word would cause the same exception to be thrown.
The problem does not occur when the document is saved as .doc.